Feature: Pasta Routes Criada
Separado resposabilidades por pastas. Cada tipo de rota tem sua própria pasta.
