+++ backend/src/routes/index.ts
@@ -1,0 +1,6 @@
+import express from 'express'
+
+const router = express.Router()
+
+export default router;
+
