+++ backend/src/server.ts
@@ -1,10 +1,6 @@
-// import firebaseAccountService from "./config/firebaseAccountService"
-
-
 import express from 'express';
 import cors from 'cors';
-import router from './routes';
-// import { config } from 'dotenv';
+import router from './routes/index';
 
 const PORT = 4000;
